backend-python/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Security & Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
PyJWT==2.10.1

# Utils & Configuration
pyyaml==6.0.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
python-multipart==0.0.7

# Development
python-dotenv==1.0.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
